requirements.txt
asgiref==3.11.1
bleach==6.3.0
dj-database-url==3.1.2
Django==6.0.4
django-taggit==6.1.0
gunicorn==25.3.0
Markdown==3.10.2
packaging==26.1
python-dotenv==1.2.2
sqlparse==0.5.5
webencodings==0.5.1
whitenoise==6.12.0
